Massive refactor
This refactor aims to move fetching of "external" details up front,
outside of any Check. This makes for a more understanding flow, IMO, and
makes Checks into pure functions which take some external details, the
dep, and possibly suggest things.

One reason for the cache feature was so that Checks that needed similar
information could naively fetch it and we'd avoid redundant work.
However, it brought some problems with it. Centralizing the calls to
external services solves this better and so I'll remove the caching
feature next commit