Fix for -r nightly
If you use nightly as your resolver we can't find "Version on this page"
because Stackage changes the label when looking at a Nightly. Checking
for and rewriting that as part of parsing the HTML was the quickest way
to get that working.

As a use-case, running with -r nightly, even if you use some LTS, will
report the extra-deps that are likely to show up eventually. This is
good when you want are looking for anything you may need to
fork/maintain to get into Stackage -- the ones reported by this can be
safely ignored.