Parse Options out of pragma comments
The contents of the linted path are "parsed" for comments of the
following format:

```yaml
\# @sled [option...]
```

Where `option...` is a valid command-line for `stack-lint-extra-deps`.

If successful, these options are merged into the CLI options (with CLI
taking precedence).

Lines without the `@sled` prefix are ignored. Lines with the prefix but
content that doesn't parse as shell-words, or does parse but are invalid
usage, are also ignored, but with warnings printed to `stderr`.

While any options can be given this way (except `--path`, which will
have no effect), the most likely use-case is lines like:

```yaml
\# @sled --exclude this-dep --exclude '**/that-dep'
```

Which allow you to centralize (and document) your reasons for excluding
certain dependencies.