Include extra-dep in GHA title
On the title is visible in the build logs, and the build logs are where
you are linked to first when you get a failure, and the annotations
aren't visible there (only summary). All of this means that you see
this:

![](https://files.pbrisbin.com/screenshots/screenshot.1651846.png)

Which gives little indication what is going on. Including the extra-dep
itself should help.