Parse trailing comments
This makes our "parser" to look for comments anywhere on the line,
rather than only on lines by themselves. It is a little less rebust:

```yaml
\# @sled <we got this before>
example1:

example2: # @sled <we'll get this now too>

example3: |
  # @sled <we *incorrectly* got this before>

example4: "xxx# @sled <we'll incorrectly get this now too>"
```

I think, given the lack of random strings as useful or valid in a
`stack.yaml`, as well as the `@sled` requirement, this is perfectly
acceptable.