Implement --fix
The main reason `--fix` has been so elusive, even once `yaml-marked` was
available and `Data.Yaml.Marked.Replace` was ready, is that, as soon as
you replace the content for one fix, that throws off the marks of
anything after it, meaning replacing those would then not do what you
expect.

To address this, I changed the main `runSLED` routine to be a loop of
parsing, suggesting, and perhaps fixing **one suggestion at a time**.

As this loop runs, we maintain the state of the underlying `ByteString`
as fixes are made. If a replacement happens, the next round would parse
from the updated contents and so the suggestions it made would have
accurate marks and be able to make a correct replacement too.

This also required tracking the (marks of) suggestions made already, so
that on each loop (in the case you _weren't_ auto-fixing) didn't
re-suggest the same things every time.

All of this got batched up into a new `Context` type and the routine is
placed in `StateT`. I chose `StateT` (instead of `MonadState`) because
we need to `lift` out to perform actions that require `MonadUnliftIO`.
If I can change that in the future, we could consider `MonadState` and
incorporate `Context` into our existing `AppT`, rather than adding a new
concrete layer.