Implement --no-check-resolver, and --checks=none
The `--checks` option's implementation is very tied to extra-deps
checking, so I'm going to leave it that way for now (instead of adding
something like `--checks=resolver`). These separate changes are a way to
be able to skip the resolver check, or do only the resolver check,
without getting into that thicket just yet.