Decode HackageExtraDep with Data.Yaml.Marked
Rather than treating it as opaque JSON. Additionally, leverage our
`ReadP` helpers to actually implement a real parse instead of the
existing textual splitting and reading.

This allows us to store a `Marked Version`, which is consistent with how
we mark only the `commit` of a git extra-dep.

It means our errors will now look like this:

    Update hspec-2.8.3 to 2.11.10
      ├ /home/patrick/code/freckle/stack-lint-extra-deps/test/examples/lts-18.18.yaml:14:5
      │    |
      │ 14 |  - hspec-2.8.3
      │    |          ^^^^^
      │
      └ A newer version is available

Instead of this:

    Update hspec-2.8.3 to 2.11.10
      ├ /home/patrick/code/freckle/stack-lint-extra-deps/test/examples/lts-18.18.yaml:14:5
      │    |
      │ 14 |  - hspec-2.8.3
      │    |  ^^^^^^^^^^^^^
      │
      └ A newer version is available

Marking specifically the version will also be required when we implement
auto-fix as a replacement of just the version bit.