Fix release workflow
I used the wrong input name.

The build still ran, but would not put the binary in the correct place
and we would've shipped and empty archive. So as part of fixing this, I
added that `--version` so that we will instead fail if a mistake that
like that happens again (or my fix is still incorrect).