chore(ci): use built binary in test-action
Previously, we had a run of the action in the `test` job that used the
built binary (`no-install`), then we had a run of the action in
`test-action` (per os) that installed the latest release.

The benefit here was that the run in `test` exercised a built binary in
the context of the action, and the runs `test-action` exercised the
install step on each os. The downside is that if there is a problem with
the currently-released version (as there is right now), the
`test-action` runs will fail. Since we have strict status checks on PRs,
this makes it difficult to address that latest release.

This commit addresses the downside while losing that upside: it makes
the installation steps become untested. I think that's acceptable
because the bulk of the logic is provided by a 3rd-party action
(`pbrisbin/setup-tool-action`) which has its own test suite. Querying
for "latest" is the only logic truly becoming untested (for now).