fix(ci): test installation only in test-action
Running SLED in the context of an action is already tested as a step in
the `test` job. `test-action` exists to test the (os-specific)
installation steps, so it installs the latest release. Sometimes (as is
the case now) there is a problem with that release which can cause
`test-action` to fail. This can make it hard to merge the PR that is
fixing things, since that status is required.

By adding a `run` input and not actually running the tool in
`test-action`, we can avoid this scenario by making the test only test
what it cares about, and so be more robust to false failures like this.